third_party/blink/web_tests/external/wpt/bluetooth/server/getPrimaryService/gen-disconnected-device.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] getPrimaryService called before connecting. Reject with NetworkError.
  assert_equals: Unexpected Error Message: expected "GATT Server is disconnected. Cannot retrieve services. (Re)connect first with `device.gatt.connect`." but got "Failed to execute 'getPrimaryService' on 'BluetoothRemoteGATTServer': GATT Server is disconnected. Cannot retrieve services. (Re)connect first with `device.gatt.connect`."
Harness: the test ran to completion.
